Proton value saving has succesfully worked
Worked on saving the proton integer value and making it usable in the Element Info class. While testing, the layout was changed according to the amount of protons (Worked as planned). We need to add some more lines of code in each element if statement though, for this to work. Tried working on saving activity state, didn't get far. Will work on that next time. In conclusion, all we need to do is add some lines of code to each if statement, and create more xml files for the elements.

-Srinath